[Bukkit] Refactor CommandExecutor out of UserIdentityLinker to WsmcBukkitPlugin
